webapi/tct-backgrounds-css3-tests/backgrounds/CSS3BG_background_f00_no-repeat_right_none.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        haox,li <haox,li@intel.com>
        Cui, Jieqiong <jieqiongx.cui@intel.com>

-->

<html>
  <head>
    <title>CSS3 Backgrounds Test: CSS3BG_background_f00_no-repeat_right_none</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/CR-css3-background-20120724/#background" />
    <meta name="assert" content="Check if 'background:#f00 none no-repeat right;' on test div"/>
    <meta name="flags" content="" />
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
    <script type="text/javascript" src="support/support.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script>
        var div = document.querySelector("#test");
        test(function () {
            div.style[headProp("background")] = "#f00 none no-repeat right";
            var backgroundColor = GetCurrentStyle("background-color");
            assert_equals(backgroundColor, "rgb(255, 0, 0)", "The background-color value");
            var backgroundImage = GetCurrentStyle("background-image");
            assert_equals(backgroundImage, "none", "The background-image value");
            var backgroundRepeat = GetCurrentStyle("background-repeat");
            assert_equals(backgroundRepeat, "no-repeat", "The background-repeat value");
            var backgroundPosition = GetCurrentStyle("background-position");
            assert_equals(backgroundPosition, "100% 50%", "The background-position value");
        }, document.title);
    </script>
  </body>
</html>
